dziala przejscie na baze

+++ silownia/database/migrations/2026_01_23_092914_add_guard_name_to_roles_table.php
@@ -10,19 +10,23 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up(): void
-    {
-       Schema::table('roles', function (Blueprint $table) {
-        $table->string('guard_name')->default('web');
+{
+    Schema::table('roles', function (Blueprint $table) {
+        if (!Schema::hasColumn('roles', 'guard_name')) {
+            $table->string('guard_name')->default('web');
+        }
     });
-    }
+}
 
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
-        Schema::table('roles', function (Blueprint $table) {
-        $table->dropColumn('guard_name');
+   public function down(): void
+{
+    Schema::table('roles', function (Blueprint $table) {
+        if (Schema::hasColumn('roles', 'guard_name')) {
+            $table->dropColumn('guard_name');
+        }
     });
-    }
+}
 };
